requirements.txt
python-nmap==0.7.1
PyYAML==6.0
requests==2.31.0
